Move get_message_id function to mailboxes
