+++ mail/libraries/mailbox_service.py
@@ -1,6 +1,4 @@
 import logging
-from email.parser import BytesHeaderParser
-from email.utils import parseaddr
 from poplib import POP3_SSL, error_proto
 from typing import Callable, Iterator, List, Tuple
 
@@ -11,63 +9,7 @@ from mail.libraries.helpers import to_mail_message_dto
 from mail.models import Mail
 from mailboxes.enums import MailReadStatuses
 from mailboxes.models import MailboxConfig, MailReadStatus
-
-
-def is_from_valid_sender(msg_header, valid_addresses):
-    parser = BytesHeaderParser()
-    header = parser.parsebytes(b"\n".join(msg_header))
-
-    logging.info("Parsing address header From: %s", header["From"])
-    name, from_address = parseaddr(str(header["From"]))
-    logging.info("Found from address %s, %s", name, from_address)
-    valid_addresses = [address.replace("From: ", "") for address in valid_addresses]
-
-    return from_address in valid_addresses
-
-
-def get_message_id(pop3_connection, listing_msg):
-    """
-    Takes a single line from pop3 LIST command and extracts
-    the message num. Uses the message number further to extract header information
-    from which the actual Message-ID is extracted.
-
-    :param pop3_connection: pop3 connection instance
-    :param listing_msg: a line returned from the pop3.list command, e.g. b"2 5353"
-    :return: the message-id and message_num extracted from the input, for the above example: b"2"
-    """
-    msg_num = listing_msg.split()[0]
-
-    # retrieves the header information
-    # 0 indicates the number of lines of message to be retrieved after the header
-    msg_header = pop3_connection.top(msg_num, 0)
-
-    if not is_from_valid_sender(msg_header[1], [settings.SPIRE_FROM_ADDRESS, settings.HMRC_TO_DIT_REPLY_ADDRESS]):
-        logging.warning(
-            "Found mail with message_num %s that is not from SPIRE (%s) or HMRC (%s), skipping ...",
-            msg_num,
-            settings.SPIRE_FROM_ADDRESS,
-            settings.HMRC_TO_DIT_REPLY_ADDRESS,
-        )
-        logging.debug("Mail was from %s", msg_header[1])
-        return None, msg_num
-
-    message_id = None
-    for index, item in enumerate(msg_header[1]):
-        hdr_item_fields = item.decode("utf-8").split(" ")
-        # message id is of the form b"Message-ID: <963d810e-c573-ef26-4ac0-151572b3524b@email-domail.co.uk>"
-
-        if len(hdr_item_fields) == 2:
-            if hdr_item_fields[0].lower() == "message-id:":
-                value = hdr_item_fields[1].replace("<", "").replace(">", "")
-                message_id = value.split("@")[0]
-        elif len(hdr_item_fields) == 1:
-            if hdr_item_fields[0].lower() == "message-id:":
-                value = msg_header[1][index + 1].decode("utf-8")
-                value = value.replace("<", "").replace(">", "").strip(" ")
-                message_id = value.split("@")[0]
-
-    logging.info("Extracted Message-Id as %s for the message_num %s", message_id, msg_num)
-    return message_id, msg_num
+from mailboxes.utils import get_message_id
 
 
 def get_read_messages(mailbox_config):
